Rip out definitions before translating
Isabelle has a horrible syntax; trying to spot the start/end of a particular
mutually recursive definition, for example, is a pain. Rather than attempting
to parse the Isabelle code, we can discard definitions from the incoming SMTLib
data, since that has a nice s-expression syntax (in fact, we assume it's one
definition per line and use grep).

Also added a bunch of extra definitions to remove; most notably custom-+, whose
termination isn't found by Isabelle, and all of the definitions which use it.