Bump te-benchmarks to avoid infinite loop
Isabelle seems to be strictly evaluating our function definitions, which is
causing it to hang when we use a function to branch between base and recursive
cases in custom-div: in a lazy language, the branch would be chosen then
evaluated, but in a strict one *both* branches are evaluated, causing infinite
recursion.

This new version of te-benchmarks avoids this problem by using pattern-matching
instead, which Isabelle does seem to evaluate lazily.