Convert extract_equations to Python
The most important change is that we buffer the entire contents of stdin as it's
read: if we see the "BEGIN OUTPUT"/"END OUTPUT" sentinels, then whatever's
between them will be formatted and emitted to stdout. If the sentinels *arent'*
found, then the whole of stdin will be sent to stderr for debugging purposes.