Replace simplistic Isabelle type parser with a more accurate one
The simplistic version was giving errors, which I've now included as a
regression test. I've also added a bunch of QuickCheck tests to ensure this
version is more robust.

In particular, we were splitting based on (, ) and =>; we gobbled up spaces as
part of the names, e.g. "(int list) list => bool" would go to "(" "int list" ")"
"=>" "bool". Now we have distinct phases:

 - Tokenising turns characters into tokens; the above would become "(" "int",
   "list" ")" "list" "=>" "bool".
 - Parenthesising turns a list of tokens into nested parenthesised trees; the
   above would become [["int", "list"], "list", "=>", "bool"].
 - Argument gathering walks along these lists, recursively, to make types. The
   above would become 'Func([Group(Group("int", "list"), "list")], "bool")'.